update the three dot menu poistioning logic
