+++ src/pages/settings/Subscription/CardSection/CardSectionActions/index.tsx
@@ -1,12 +1,9 @@
-import React, {useCallback, useMemo, useRef} from 'react';
-import {View} from 'react-native';
+import React, {useMemo} from 'react';
 import * as Expensicons from '@components/Icon/Expensicons';
 import ThreeDotsMenu from '@components/ThreeDotsMenu';
 import type ThreeDotsMenuProps from '@components/ThreeDotsMenu/types';
 import useLocalize from '@hooks/useLocalize';
-import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import Navigation from '@navigation/Navigation';
-import type {AnchorPosition} from '@styles/index';
 import CONST from '@src/CONST';
 import ROUTES from '@src/ROUTES';
 
@@ -16,9 +13,7 @@ const anchorAlignment = {
 };
 
 function CardSectionActions() {
-    const {shouldUseNarrowLayout} = useResponsiveLayout();
     const {translate} = useLocalize();
-    const threeDotsMenuContainerRef = useRef<View>(null);
 
     const overflowMenu: ThreeDotsMenuProps['menuItems'] = useMemo(
         () => [
@@ -36,29 +31,13 @@ function CardSectionActions() {
         [translate],
     );
 
-    const calculateAndSetThreeDotsMenuPosition = useCallback(() => {
-        if (shouldUseNarrowLayout) {
-            return Promise.resolve({horizontal: 0, vertical: 0});
-        }
-        return new Promise<AnchorPosition>((resolve) => {
-            threeDotsMenuContainerRef.current?.measureInWindow((x, y, width, height) => {
-                resolve({
-                    horizontal: x + width,
-                    vertical: y + height,
-                });
-            });
-        });
-    }, [shouldUseNarrowLayout]);
-
     return (
-        <View ref={threeDotsMenuContainerRef}>
-            <ThreeDotsMenu
-                getAnchorPosition={calculateAndSetThreeDotsMenuPosition}
-                menuItems={overflowMenu}
-                anchorAlignment={anchorAlignment}
-                shouldOverlay
-            />
-        </View>
+        <ThreeDotsMenu
+            shouldSelfPosition
+            menuItems={overflowMenu}
+            anchorAlignment={anchorAlignment}
+            shouldOverlay
+        />
     );
 }
