+++ src/pages/settings/Subscription/TaxExemptActions/index.tsx
@@ -1,12 +1,10 @@
-import React, {useCallback, useMemo, useRef} from 'react';
+import React, {useMemo} from 'react';
 import {View} from 'react-native';
 import * as Expensicons from '@components/Icon/Expensicons';
 import ThreeDotsMenu from '@components/ThreeDotsMenu';
 import type ThreeDotsMenuProps from '@components/ThreeDotsMenu/types';
 import useLocalize from '@hooks/useLocalize';
-import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-import type {AnchorPosition} from '@styles/index';
 import {navigateToConciergeChat} from '@userActions/Report';
 import {requestTaxExempt} from '@userActions/Subscription';
 import CONST from '@src/CONST';
@@ -17,10 +15,8 @@ const anchorAlignment = {
 };
 
 function TaxExemptActions() {
-    const {shouldUseNarrowLayout} = useResponsiveLayout();
     const styles = useThemeStyles();
     const {translate} = useLocalize();
-    const threeDotsMenuContainerRef = useRef<View>(null);
 
     const overflowMenu: ThreeDotsMenuProps['menuItems'] = useMemo(
         () => [
@@ -37,27 +33,10 @@ function TaxExemptActions() {
         [translate],
     );
 
-    const calculateAndSetThreeDotsMenuPosition = useCallback(() => {
-        if (shouldUseNarrowLayout) {
-            return Promise.resolve({horizontal: 0, vertical: 0});
-        }
-        return new Promise<AnchorPosition>((resolve) => {
-            threeDotsMenuContainerRef.current?.measureInWindow((x, y, width, height) => {
-                resolve({
-                    horizontal: x + width,
-                    vertical: y + height,
-                });
-            });
-        });
-    }, [shouldUseNarrowLayout]);
-
     return (
-        <View
-            ref={threeDotsMenuContainerRef}
-            style={[styles.mtn2, styles.pAbsolute, styles.rn3]}
-        >
+        <View style={[styles.mtn2, styles.pAbsolute, styles.rn3]}>
             <ThreeDotsMenu
-                getAnchorPosition={calculateAndSetThreeDotsMenuPosition}
+                shouldSelfPosition
                 menuItems={overflowMenu}
                 anchorAlignment={anchorAlignment}
                 shouldOverlay
